Remodeled the data obtaining into various different classes.
