+++ crypto_history/__init__.py
@@ -1,4 +1,5 @@
 from .utilities.general_utilities import init_logger, AbstractFactory
-from .core import get_market_data, data_container_pre
+from .core import get_market_data, data_container_pre_1
+from .core import data_container_pre
 
 class_builders = AbstractFactory.get_builders()
